Move mpfit to extern subpackage
