+++ docs/code_reference/eispac_core.rst
@@ -1,5 +1,9 @@
 eispac core
 ===========
 
+All functions and classes in ``core`` are loaded into the top-level ``eispac``
+namespace. It is therefore recommended to access them via calls like
+``eispac.FUNCTION()`` rather than ``eispac.core.FUNCTION()``
+
 .. automodapi:: eispac.core
    :no-heading:
