+++ docs/code_reference/eispac_extern.rst
@@ -1,0 +1,8 @@
+eispac extern
+=============
+
+This subpackage contains third-party code that is included in ``eispac`` for
+convenience or efficiency.
+
+.. automodapi:: eispac.extern
+  :no-heading:
