+++ eispac/__init__.py
@@ -1,7 +1,8 @@
-__version__ = '0.94.0'
+__version__ = '0.94.1'
 
 from . import download
 from . import templates
 from .core import *
+from .extern import *
 from .util import *
 from . import data
