+++ eispac/core/__init__.py
@@ -1,7 +1,6 @@
 from .eiscube import *
 from .eisfitresult import *
 from .eismap import *
-from .mpfit import *
 from .read_wininfo import *
 from .read_cube import *
 from .read_fit import *
