+++ eispac/core/fit_spectra_astropy.py
@@ -9,7 +9,7 @@ from eispac.core.eiscube import EISCube
 from eispac.core.read_template import EISFitTemplate
 from eispac.core.eisfitresult import EISFitResult
 from eispac.core.scale_guess import scale_guess
-from eispac.core.mpfit import mpfit
+from eispac.extern.mpfit import mpfit
 from eispac.core.generate_astropy_model import generate_astropy_model
 
 def fit_spectra_astropy(inten, template, parinfo=None, wave=None, errs=None,
