+++ eispac/extern/README.md
@@ -1,0 +1,5 @@
+# External Code Subpackage
+
+This subpackage contains third-party code that is included in `eispac` for
+convenience or efficiency (currently only `mpfit.py`). Any programs added to
+this subpackage _must_ have a licence that permits redistibuting the code.
