+++ eispac/extern/__init__.py
@@ -1,0 +1,1 @@
+from .mpfit import *
