+++ eispac/net/tests/test_eis_client.py
@@ -13,7 +13,7 @@ def eis_query():
     level = a.Level('1')
     return time, instr, obs, source, provider, level
 
-@pytest.mark.skip(reason="NRL server maintenance")
+# @pytest.mark.skip(reason="NRL server maintenance")
 @pytest.mark.remote_data
 def test_search_all_types(eis_query):
     q = Fido.search(*eis_query)
@@ -21,7 +21,7 @@ def test_search_all_types(eis_query):
     assert len(q[0]) == 3
     assert q[0,0]['url'] == 'https://eis.nrl.navy.mil/level1/hdf5/2022/03/29/eis_20220329_222113.data.h5'
 
-@pytest.mark.skip(reason="NRL server maintenance")
+# @pytest.mark.skip(reason="NRL server maintenance")
 @pytest.mark.remote_data
 def test_search_fits_only(eis_query):
     q = Fido.search(*eis_query, a.eispac.FileType('FITS'))
@@ -29,14 +29,14 @@ def test_search_fits_only(eis_query):
     assert len(q[0]) == 1
     assert q[0,0]['url'] == 'https://eis.nrl.navy.mil/level1/fits/2022/03/29/eis_er_20220329_222113.fits'
 
-@pytest.mark.skip(reason="NRL server maintenance")
+# @pytest.mark.skip(reason="NRL server maintenance")
 @pytest.mark.parametrize('file_type, file_url', [
     ('FITS', 'https://eis.nrl.navy.mil/level1/fits/2022/03/29/eis_er_20220329_222113.fits'),
     ('HDF5 data', 'https://eis.nrl.navy.mil/level1/hdf5/2022/03/29/eis_20220329_222113.data.h5'),
     ('HDF5 header', 'https://eis.nrl.navy.mil/level1/hdf5/2022/03/29/eis_20220329_222113.head.h5')
 ])
 
-@pytest.mark.skip(reason="NRL server maintenance")
+# @pytest.mark.skip(reason="NRL server maintenance")
 @pytest.mark.remote_data
 def test_search_individual_filetypes(eis_query, file_type, file_url):
     q = Fido.search(*eis_query, a.eispac.FileType(file_type))
@@ -45,7 +45,7 @@ def test_search_individual_filetypes(eis_query, file_type, file_url):
     assert q[0,0]['url'] == file_url
     assert q[0,0]['FileType'] == file_type
 
-@pytest.mark.skip(reason="NRL server maintenance")
+# @pytest.mark.skip(reason="NRL server maintenance")
 @pytest.mark.remote_data
 def test_combined_hdf5_search(eis_query):
     q = Fido.search(*eis_query,
